Update OM_2b_Update_catch.R
Corrections where effort at age is transposed incorrectly when using multiple iterations.  The effort per iteration vector was being recycled within a single iteration across ages, causing incorrect behaviour.

This works as it repeats the iteration effort for each age within an iteration.  However, it may not be an ideal solution, and it may be better to correct the transposition so there is no recycling of values....?