Bug in the interaction btw fleets and stocks in BD populations
FLBEIA was computing the catch and effort against the bomass in the start of the season (effort), or even worst (catch) agains the biomass at the start reduced by an exponential growth (exp(-M/2)). Now it has been corrected and in all cases in uses the biomass plus the growth. It has been changse in all the fleet dynamic model functions but SSFB.