SimultaneousMngt argument added to main.ctrl
This argument was added to be able to have simultaneous management to
used multistock HCRs. As it was the whole MP was run stock by stock.
SimultaneousMngt = FALSE  corresponds with the original configuration
and it is filled at the start of FLBEIA if the argument is missing in
main.ctrl.

Some other changes to fix problems with biomass dynamic populations.