@index.var removed from ageInd/bioInd, @sdb added to FLBDsim and and calculate.q.sel.flrObjs added
o Due to incompatibility with the use of .@index.var in stock assessment models (sca)
index.var is no longer used to store the uncertainty of the index. The uncertainty can be incorporate in
the slot index.q, ie., @index.q = q*err

o  @sdb standard deviation of the process error in a continuos prodyction model (SPiCT). It adds the term -0.5*sdb*Delta_t to the stock dynamics equation.