New Function setUnitsNA
It sets all the units in FLBiol or FLFleet objects to NA. In some cases having units produce and error in 'sprintf' that we don't know how to solve.

fltSum and fltStkSum, the error appears in these two functions, hence as the result of this functions do not use units, we first remove the units from the objects and then proceed.